refactor(database): simplify database connection and remove godotenv dependency
- Remove godotenv dependency and related code as environment variables are now used directly
- Change StockPrice type from int64 to float64 in models for better precision
- Clean up error messages and improve code consistency in middlewares
- Update go.mod and go.sum to reflect dependency changes

+++ models/models.go
@@ -3,6 +3,6 @@ package models
 type Stock struct {
 	StockID      int64  `json:"stockid"`
 	StockName    string `json:"stockname"`
-	StockPrice   int64  `json:"stockprice"`
+	StockPrice   float64  `json:"stockprice"`
 	StockCompany string `json:"stockcompany"`
 }
